+++ lib/getch/filesystem/ext4/partition.rb
@@ -12,8 +12,8 @@ module Getch
 
         def run_partition
           return if STATES[:partition ]
-          @clean.hdd(@disk, @cache_disk, @home_disk)
-          @clean.boot(@boot_disk, @disk)
+          @clean.hdd(@disk)
+          @clean.external_disk(@disk, @boot_disk, @cache_disk, @home_disk)
           if Helpers::efi?
             partition_efi
           else