+++ lib/getch/guard.rb
@@ -4,6 +4,12 @@ end
 class InvalidFormat < StandardError
 end
 
+class InvalidZone < StandardError
+end
+
+class InvalidKeymap < StandardError
+end
+
 module Getch::Guard
   def self.disk(name)
     raise InvalidDisk, "No disk." unless name
@@ -19,6 +25,35 @@ module Getch::Guard
     raise InvalidFormat, "Format #{name} not yet available." if name.match(/btrfs/)
     raise InvalidFormat, "Format #{name} not supported." unless name.match(/zfs|lvm|zfs/)
   rescue InvalidFormat => e
+    puts "#{e.class} => #{e}"
+    exit 1
+  end
+
+  def self.zone(name)
+    raise InvalidZone, "No zoneinfo specified." unless name
+    raise InvalidZone, "Directory /usr/share/zoneinfo/ no found on this system..." unless Dir.exist? "/usr/share/zoneinfo/"
+    raise InvalidZone, "Zoneinfo #{name} is no found in /usr/share/zoneinfo/." unless File.exist? "/usr/share/zoneinfo/#{name}"
+  rescue InvalidZone => e
+    puts "#{e.class} => #{e}"
+    exit 1
+  end
+
+  def self.keymap(name)
+    raise InvalidKeymap, "No keymap specified." unless name
+    key = []
+
+    if Dir.exist? '/usr/share/keymaps'
+      key = Dir.glob("/usr/share/keymaps/**/#{name}.map.gz")
+    elsif Dir.exist? '/usr/share/kbd/keymaps'
+      key = Dir.glob("/usr/share/kbd/keymaps/**/#{name}.map.gz")
+    else
+      raise InvalidKeymap, "No directory found for keymap."
+    end
+
+    if key == []
+      raise InvalidKeymap, "Keymap #{name} no found."
+    end
+  rescue InvalidKeymap => e
     puts "#{e.class} => #{e}"
     exit 1
   end