+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-* Control input for `--disk DISK`, `--format FS`.
-* Add a `/etc/portage/bashrc` to automatically signing kernel modules with `Portage`.
-* Now clean properly another disk used with option s`--separate-x DISK`.
+* Control input for options `--disk`, `--format`, `--zoneinfo`, `--keymap`.
+* Add a `/etc/portage/bashrc` to automatically signing kernel modules with `emerge`.
+* Now clean properly another disk used with option `--separate-x DISK`.
 * For ZFS, use `blockdev --getpbsz` to find correct bloc (sector) size.
 * Populate `/etc/modules-load.d/` with modules found with `lsmod` (only few wifi's, flash usb related (ehci, ohci, xhci, etc)).
 * Remove the package `dev-util/dwarves`.