+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Install `iptables` by default.
 * Control input for options `--disk`, `--format`, `--zoneinfo`, `--keymap`.
 * Add a `/etc/portage/bashrc` to automatically signing kernel modules with `emerge`.
 * Now clean properly another disk used with option `--separate-x DISK`.